rename search_user_items to search_users_items
