+++ server/controllers/items/search.js
@@ -1,4 +1,4 @@
-const searchUserItems = require('./lib/search_user_items')
+const searchUsersItems = require('./lib/search_users_items')
 const { filterPrivateAttributes } = require('controllers/items/lib/filter_private_attributes')
 const { getAllowedVisibilityKeys } = require('lib/visibility/allowed_visibility_keys')
 
@@ -9,7 +9,7 @@ const sanitization = {
 
 const controller = async ({ reqUserId, userId, search }) => {
   const allowedVisibilityKeys = await getAllowedVisibilityKeys(userId, reqUserId)
-  const items = await searchUserItems({ search, userId, reqUserId, allowedVisibilityKeys })
+  const items = await searchUsersItems({ search, userId, reqUserId, allowedVisibilityKeys })
   return {
     items: items.map(filterPrivateAttributes(reqUserId))
   }
